requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
flask-oidc
oauth2client
config
gunicorn
psycopg2
pytest